backend/requirements.txt
"""FastAPI Backend Requirements"""
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0

# Firebase
firebase-admin==6.2.0

# Database
python-dotenv==1.0.0

# JWT & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.25.0
aiohttp==3.9.0

# Logging & Monitoring
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9

# CORS & Middleware
python-multipart==0.0.6

# Optional: Caching
redis==5.0.0

# Optional: Async ORM (if using PostgreSQL later)
sqlalchemy==2.0.23
asyncpg==0.29.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
